main_ai/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
requests>=2.28.0
beautifulsoup4>=4.11.0
numpy>=1.21.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6